--- src/bus/components/Editor.vue
@@ -1,22 +1,0 @@
-<template>
-  <div id="editor">
-    <h1>エディター画面</h1>
-    <span>{{ user.displayName }}</span>
-    <button @click="logout">ログアウト</button>
-  </div>
-</template>
-
-<script>
-  export default {
-    name: "editor",
-    props: ["user"],
-    data() {
-      return {};
-    },
-    methods: {
-      logout: function() {
-        firebase.auth().signOut();
-      }
-    }
-  };
-</script>
